WL#7802 PERFORMANCE SCHEMA, BATCH TABLE IO
Description:
------------
This WL will reduce number of calls for PFS when MySQL do a single table scan 
or join operation.  
We set batch mode for PFS when do join operation in executor to implement
reduce number of calls for PFS.
This WL does not change any functionality of join operation.

Notice:
This version covers:
1 Simplify the code by Roy's suggestion.

Remaining problem:
1 A bug that make the server crash.
  It occurs in Protocol::end_statement()->DBUG_ASSERT(0). For example:
    ./mtr i_main.bka
  It need to be resolved in next version. 
  Analyzing...