WL#8007 - GCS: Server Interfaces
(Transaction source  post fix)

On server's before_commit hook invocation a parameter was passed to
identify the  transaction origin.

This condition is however to restrict, not allowing the certification
that are originated in the server SQL thread in the case that it is
receiving events trough a slave connection.

To solve this, the test is now done in the plugin based on the thread
id of the applying thread.