BUG#17787664 - ASSERT `! IS_SET()` IN DIAGNOSTICS_AREA::SET_OK_STATUS                ON ER_SUBQUERY_NO_1_ROW
ANALYSIS :

The SELECT subquery in the below query triggers an assert on
a debug build.

DELETE  FROM OUTR1.* USING t1 AS OUTR1
    LEFT OUTER JOIN t2 AS OUTR2  ON ( OUTR1.a = OUTR2.b )
WHERE OUTR1.a < ( SELECT t3.c  from t3 where 1 XOR OUTR2.b IS NOT NULL);

SELECT subquery returns two records which is an incorrect
input for the comparison operator('<'). Hence the execution of
the query sets the status of the diagnostic area to DA_ERROR
due to the error 'ER_SUBQUERY_NO_1_ROW'. The query
execution continues due to a missing error check in the
function 'evaluate_null_complemented_join_record()'.
While trying to set the status of the diagnostic area to
DA_OK, the assert is triggered since the error condition
is already set in the diagnostic area for the query.

In case of a release build, an appropriate error
'ER_SUBQUERY_NO_1_ROW' is reported.

FIX:

Added error check in evaluate_null_complemented_join_record().