Bug#18095449 REPLICATE_WILD_DO_TABLE AND REPLICATE_WILD_IGNORE_TABLE ACCEPTING INVALID VALUES
Problem: REPLICATE_WILD_DO_TABLE AND REPLICATE_WILD_IGNORE_TABLE
filter values should be in the format of 'db_pattern.table_pattern'.
Invalid values (value without a dot ('.') in it) are also getting accepted
which is not correct.

Analysis: A dot is used as a separator in the table filter list values.
If a filter value is found without a '.' in it, it is ignored and 
will not be accepted. But in the server, before the logic reaches 
to this check, Server is clearning the existing values. In this situation,
the existing (old) values are getting cleared and new values are not
getting set.
Eg: Replicate_wild_do_table= db1.t%
If Change replication filter Replicate_wild_do_table=('t2%') is executed
db1.t% is cleared and 't2%' is not accepted and also the command is
not throwing any error.

Fix: Change the Wild_[do|ignore]_table syntax at parser level
not to allow the pattern if it does not contain a dot(.).
Now we are introducing new tokenizer "qualified_table_as_string"
which will treat any string without 'dot' as invalid and will
throw error "ER_INVALID_RPL_WILD_TABLE_FILTER_PATTERN".