Bug#19233510 [ERROR] INNODB: IGNORING THE REDO LOG DUE TO MISSING MLOG_CHECKPOINT
InnoDB crash recovery expects the redo log to be empty after the
latest checkpoint, or to contain a MLOG_CHECKPOINT marker.

Sometimes, a MLOG_CHECKPOINT marker was incorrectly being omitted when
performing a log checkpoint.

We must only omit the marker when performing a shutdown, to avoid
hitting an infinite loop.

This bug used to mostly occur in tests that invoke extra checkpoints,
such as those related to WL#6501 (TRUNCATE TABLE) and WL#7277
(bulk load for index creation).

rb#6330 approved by Jimmy Yang