Fix Bug#18663997 INNODB: TOO MANY FILES STAY OPEN
Flushing is not done for FIL_TYPE_TEMPORARY files (at the beginning
of fil_flush()):

        if (space == NULL
            || space->purpose == FIL_TYPE_TEMPORARY
            || space->stop_new_ops
            || space->is_being_truncated) {
                mutex_exit(&fil_system->mutex);

                return;
        }

but FIL_TYPE_TEMPORARY and FIL_TYPE_IMPORT were considered equivalent
and thus flushing was also not done for FIL_TYPE_IMPORT.

This lead to the problem that modification counter > 0 and
flush counter == 0 for to-be-imported tablespaces and it was left like
that after the tablespace got converted to normal FIL_TYPE_TABLESPACE
in fil_space_set_imported().

Thus, when flushing do not skip the to-be-imported tablespaces.

Approved by:	Marko (rb:6318)