Bug#18875137: NDB_RESTORE STILL MISSING SOME TYPE CONVERSIONS
Added TEXT->TEXT staging conversions to ndb_restore. Fixed ndb_restore
to handle a combination of built-in and staging conversions in a
single table. Disabled conversion to TEXT column with different
charset. Added fix in ndb_restore for truncation when converting to
TINYTEXT. Added fix in mysqld to truncate TEXT values from ndb
according to MySQL limits on TEXT sizes.

While truncating a multi-byte charset TEXT, there is a possibility
that the TEXT could be trusacated in the middle of a multi-byte
character. This is handled by truncating to the end of the last
well-formed character.