Bug#19311260 ASSERTION IN MTS WHEN WORKER IS KILLED WHILE WAITING FOR              COMMIT LOCK
Problem:
=======

With MTS enabled, when a worker is killed while waiting for commit
lock, it is causing a debug assert to fail.

Analysis:
========

When an error is reported through my_error() function call, the assert
expects 'thd->get_stmt_da()->m_status' diagnostic variable to be
'DA_EMPTY', but  when a worker reaches 
Xid_log_event::do_apply_event_worker, thd->get_stmt_da()->m_status has
'DA_OK' as value.

In STS version of Xid_log_event applier, because of the above
explained, the THD object needs to be reset so that the errors, flags
and diagnostic variables that are set by the previous command are
cleared.

Fix:
===

Added code into MTS version of Xid_log_event "do_apply_event_worker" to
reset THD for the next command the same way it is done for STS version
"do_apply_event".