Bug#19419463 ASSERT FROM->M_BYTE_COUNT == 0 AT SQL_DIGEST_STORAGE::COPY IN SQL_DIGEST.H
Before this bug, a select from table
performance_schema.events_statements_current
could cause a failed assert in debug mode.

The assert did fail under load, when reading the digest of a query
that is concurrently starting.

Note that tables
- performance_schema.events_statements_history
- performance_schema.events_statements_history_long
are not affected by this bug, as only reading from a live query
can expose the race condition found.

The root cause is that the copy is doing a dirty read by design,
but the code is not robust to handle data changing unexpectedly.

Fixed sql_digest_storage::copy to be more robust.