Bug#18477009 - INACCURATE HANDLING OF SRV_ACTIVITY_COUNT
We call srv_active_wake_master_thread() directly and one of the places is
innobase_commit (and prepare as well). This call not only wakes up the
master thread but also increments the srv_activity_count which tells
the page_cleaner that server is not idle. That's no what we expect.

We should call srv_active_wake_master_thread() only after the commitment
of a write trx, but not read-only trx, or after a rollback. This patch also
changes some call of srv_active_wake_master_thread() to
ib_wake_master_thread().

Original patch is provided by Inaam.

rb#5909, approved by Jimmy.