Bug#18509896: 0.00000000042 MOD 10 = 0.42000000000
Problem:
Mod operation on a decimal with leading zeroes gives
wrong results.
      
Analysis:
In do_div_mod while calculting the destination buffer's length 
of integer part, precision is presumed to be more than
the length of fraction always. This presumption is proven false
when leading zeroes are present in the fractional part as 
precision is calculated after stripping the leading zeroes.
      
As a result, when intg0 should have a negative value so that
the destination buffer would be filled with zeroes later, it is
right now having a positive value. Hence the change in result.
      
Solution:
Calculate the intg0 taking into consideration that prec1
can be less than frac1.