Bug #18354165 	16723708 CHANGED EVENT CATEGORY VALUES
This fix reverts the fix for 16723708 INVALID VALUES RETURNED IN NDB_LOGEVENT "CATEGORY" MEMBER VARIABLE.

The problem was that changing the category returned by ndb_logevent_get_next()
was non backwards compatible, and some customers were already working around 
the problem in their applications.

So in this fix we keep the old behaviour of ndb_logevent_get_next(), and
create a new ndb_logevent_get_next2() method which will return a 
correct category.

Customers requiring the correct category (and those not caring)
should call ndb_logevent_get_next2().

Existing customer applications will be unaffected.

There does not appear to be any internal use of the category field
returned this way which is affected.