Bug#19274972: RESULT DIFF SEEN FOR QUERIES WITH DISTINCT AND ORDER BY It's a bug introduced by wl6711 on unique_constraint implementation.
When unique_constriant is used, Hash value is treated to compare whether
two tuples are equal or not. However, Hash function can't assure the
uniqueness for different tuples. So we need to compare the content if
there is the same Hash value. During compare the content, there is an
error for comparing length in function cmp_field_value.