Followup to vasil.dimov@oracle.com-20140819130324-popsb4fncjc4db71:
Re-throw the exception if an exception is thrown from some of the
constructor calls when creating an array of objects in
ut_allocator::new_array(). That is the standard behavior of:

	foo_t*	foo = new foo_t[n];
	or
	foo_t*	foo = new(std::nothrow) foo_t[n];

(std::nothrow prevents an exception only if the raw allocation of the memory
chunk fails, not if some of the constructor calls throws an exception).