Bug #19471516 SERVER CRASHES WHEN EXECUTING ALTER TABLE ADD FOREIGN KEY
Problem:

We maintain foreign keys sorted by foreign->id in an std::set object.  If
the foreign->id is to be changed, then the foreign key object must be 
removed from std::set and re-inserted.  This is necessary to maintain the
search order in the std::set object.

The root cause of the problem was identified that the search order was 
lost in the referenced_set.  This is because while renaming the table,
we didn't not refresh this referenced_set.

Solution:

When a foreign key is renamed, we must delete and re-insert into both
foreign_set and referenced_set.  

rb#6420 approved by Jimmy.