Bug#18192817: IGNORE_SERVER_IDS IS SHOWING DUPLICATE               SERVER ID FOR LAST VALUE
Problem:

The list of insert_id is stored in an array.
Before inserting in the array, there is a binary
search to find if the same server_id already 
exists in the list. The issue is that the list
is not sorted. Note that binary search can ONLY
be applied on a sorted list.

As a result the list can sometimes contain a value
but not be spotted by the binary search logic.
In such a case, the value is again added to the list.

Fix:

The bug was fixed by the patch for Bug#18920203.
The patch for this bug implemented insertion-sort
in Prealloced_array, and used insert_unique() to
keep the array sorted at all times.

The current patch only adds a test to verify that
we dont have duplicate values in the list at any time.