Bug#17833261 FOUND_ROWS(): DIFFERENT RESULTS FOR LOCAL LIMIT WITH AND WITHOUT OFFSET IN UNION
Problem: A UNION ALL query with SQL_CALC_FOUND_ROWS and a limit with
offset for one query block reports the wrong number of found rows.

In select_union_direct::send_eof(), the number of rows is aggregated
by reading THD::limit_found_rows inc after each query block
execution. However, this number is the sum of the limit (or number of
rows, if smaller than the limit) and the offset. This number is
correct for non-union queries, but in union queries the actual number
of rows returned to the user is needed.

Fix: Subtract the offset from limit_found_rows in
select_union_direct::send_eof().