WL#8003: Server support for attachable transactions.
The patch introduces the server-side support (as opposed to the InnoDB-side
support) of attachable transactions. Attachable transaction is a tiny
subclass of a nested transaction. Attachable transactions are AC-RO-RC-NL
transactions (auto-commit, read-only, read-committed, non-locking).

A new flag (HA_ATTACHABLE_TRX_COMPATIBLE) is added to the Handler API.
Briefly, HA_ATTACHABLE_TRX_COMPATIBLE means that a SE is compatible with
the attachable transaction requirements. That in turn means that the tables
from the SE can be used within an attachable transaction. It does NOT mean
that the SE supports attachable transactions per say.

New public operations are introduced to THD to begin and end attachable
transaction.

Attachable transactions will be used later by to read from transactional
system tables.