WL#7828: InnoDB: attachable transaction support.
This is a cumulative patch from Satya Bodapati and Alexander Nozdrin.

The patch introduces the InnoDB-side support (as opposed to the server-side
support) of attachable transactions. Attachable transaction is a tiny
subclass of a nested transaction. Attachable transactions are AC-RO-RC-NL
transactions (auto-commit, read-only, read-committed, non-locking).

The patch does the following:
  - Adds asserts to verify that the attachable transaction:
    - doesn't take locks,
    - is READ-COMMITTED,
    - is READ-ONLY,
    - is AUTOCOMMIT.
  - Adds test cases to ensure that InnoDB handles attachable
    transactions properly.