Bug #18066518: THE COST VALUE IS A NEGATIVE NUMBER FOR MERGE ENGINE TABLE Bug #18486293: ASSERTION FAILED: KEYS >= 0.0 IN                COST_MODEL_TABLE::KEY_COMPARE_COST
Description:
------------
The cost numbers should not be negative.
The cost calculation code for loose index scan in cost_group_min_max()
does not handle the case where the table is empty.
When table_records == 0, the log(0) produces -INFINITY.
So we should avoid calling log(0).

Fix:
----
When MySQL calculate the cost, we ban it to be a negative number.
Fixed it by avoiding use log(0) when table_records == 0.

Test case added.