Bug#19202654  NDB_RESTORE NOT ABLE TO RESTORE META WITH 17K OBJECTS
Fixed autotest failure of test case added for this bug.
The --error-insert option is supported by ndb_restore only in 
debug builds. The test failed because this option was passed to 
ndb_restore even though it was not a debug build. Modified the 
test code to pass '--error-insert' as an option to ndb_restore 
only if the build is a debug build.