Bug#19514950 CLEAN UP TESTS THAT KILL OR RESTART THE SERVER Bug#19505636 INNODB.LOG_FILE_SIZE FAILS WITH WRONG ERRNO 2006
There are a number of tests that are killing or restarting the server.
We should try to use common infrastructure for these.

SET DEBUG='d,crash_commit_before';
should not be used for just killing the server.
A normal kill will run on a non-debug binary as well.

Valgrind should not be disabled when the server
is being killed by the test script.

There still seems to be a problem with DBUG_SUICIDE and Valgrind,
so we will have to disable Valgrind when the server is killed by
test instrumentation.

We introduce some helper files:

include/kill_mysqld.inc
include/kill_and_restart_mysqld.inc

We introduce $restart_parameters for the above, and for the existing

include/start_mysqld.inc
include/restart_mysqld.inc

"Homebrew" restart has been replaced with the helpers as much as possible.

Most use of shutdown_server has been replaced with use of one of the
include files.

rb#6485 approved by Satya Bodapati