Bug#19537790 INNODB_STATS_FETCH FAILS WITH WRONG CARDINALITY              ESTIMATE ON 32 BIT PLATFORMS
The innodb_stats_fetch test failed when running
on 32 bit platforms due to an "off by one" cardinality number
in the result from a query that read from the statitics table
in information schema. This test failure is caused by WL#7339.

The cardinality numbers that are retrieved from information schema
is roughly calculated this way when the table is stored in InnoDB:

1. InnoDB has the number of rows and the cardinality for the 
   index stored in the persistent statistics. In the failing
   case, InnoDB had 1000 as the number of rows and 3 as the 
   cardinality.
2. InnoDB calculates the records per key value and stores this
   in the KEY object. This is calculated as 1000/3 and is thus
   333.333333.... when using the code from WL#7339 (before
   this worklog, the rec_per_key value was 166).
3. When filling data into information schema, we re-calculate
   the cardinality number by using the records per key
   information (in sql_show.cc):

      double records= (show_table->file->stats.records /
                       key->records_per_key(j));
      table->field[9]->store((longlong) records, TRUE);

   in this case we first compute the cardinality to be
 
      records= 1000 / 333.3333... = 3.0000... or 2.9999999

   and then use the cast to get an integer value. On 64-bit
   platforms, the result of this was 3.00000 which was 
   casted to 3. On 32 bit platforms, the result was 2.999999
   which was casted to 2 before inserting it into the
   information schema table. (in the pre-wl7339 version, the
   calculated cardinality number was 6 for this case).

This issue is caused by having converted to using a float numbers
for the records per key estimate. So when re-calculating the
cardinatily number in step 3 above, we can easily get a
result that is just below the actual correct cardinality number
and due to the use of cast, the number is always truncated.

The suggested fix for this problem is to round the calculated
cardinality number to the nearest integer value before 
inserting it into the statistics table. This will both avoid the
issue with different results on different platforms but it will
also produce a more correct cardinality estimate.

This change has caused a few other test result files to be
re-recorded. The updated cardinality numbers are more correct
than the previous.