BUG#17069285 SEGMENTATION FAULT IN NDB_PRINT_FILE
The Bug was due to m_out object used in NdbOut operator was not initialized/set to the output stream.
Added an ndb_init() call to initialize it to output stream.
Added an environmental variable $NDB_PRINT_FILE to give path to its binary in  mysql-test/mysql-test-run.pl file.
Added a unit test and its result file.
Missed a line in previous push, so pushing again.