Bug#17896265: PARTITIONED TABLE HAS MISPLACED ROWS,               AFTER INPLACE ALTER
PROBLEM
-------
When we have multiple columns in KEY partitioning,
the order of columns is important to determine the 
partition in which the row will be placed.Altering 
the order of columns using INPLACE algorithm
can lead to inconsistency in placement of rows in 
partitions. (i.e A row inserted before ALTER will 
be placed in one partition and the same row inserted
after ALTER will be placed in different partition)
  
FIX
---

We will not allow the ALTER operation to succeed if
changing of the order causes inconsistency in 
placement of rows in partitions.