BUG#19535945 FAILED CREATE PARTITIONED TABLE LEAKS TABLE INSTRUMENTATION
Before this fix, particular illegal CREATE TABLE statements would:
- fail to create the table, as expected
- and yet still create the table statistics in the performance schema

This in the long term causes instrumentation leaks,
consuming space in performance schema buffers for tables
that do not exist.

The problem is located in ha_create_table(),
where one code path:
- creates instrumentation with PSI_TABLE_CALL(get_table_share)
- creates the table itself in the storage engine
- fails to cleanup on failures, with PSI_TABLE_CALL(drop_table_share).

The fix adds the missing call to
  PSI_TABLE_CALL(drop_table_share)
for this error path.