Bug#19047527 SQRT() : UNINITIALISED VALUE IN MY_STRTOD
We were doing an out-of-bounds read when parsing "0E+"
Visible only in valgrind run of an optimized build.
Note that the compiler is free to re-arrange the evaluation order of
(s < end && c >= '0' && c <= '9')
as long as program semantics are kept.
That's why we got a valgrind warning when testing (c >= '0')