Fix for Bug#19552283 :
  Transporters receiveBuffers might be reset while 
  another thread read them (race)

There is a potential race between a thread 
receiving data with TransporterRegistry::performReceive,
while another thread (async) initiate disconnect of the
transporter by calling Transporter::doDisconnect(). The
later method will call TCP_Transporter::disconnectImpl()
which clears the receive buffer.

A typical case where this happens is ::doDisconnect() being
called from the thread running TransporterRegistry::start_clients_thread().
Simultaneously another thread might 'do_poll()' which performReceive()
on the same transporter.

This fix moves 'receiveBuffer.clear()' out of ::disconnectImpl()
and into a new method Transporter::resetBuffers() which is now
called from ::do_connect(). At this point The transporter has
been trough a report_disconnect() synched with performReceive().
Furthermore, it is known to be in a DISCONNECTED state, which
implies that it can't be seen as connected again by 
::performReceive() as this also has to be synched.

This also introduce a relaxing of in which states buffered
data can be read & unpacked from the transporters: Any received
data will now be available in the receivebuffers as long as 
the transporter as available in 'm_has_data_transporters'.
The patch takes advantage of this by simplifying the check
of connection states in ::performReceive().

The patch also removes the 'virtual' decl. of 
Transporter::doDisconnect(). This was a leftover from the 
time before (the virtual) Transporter::disconnectImpl()
was introduced.

Lots of comments are added in order to better explain
the concurrency controll and restrictions between
::pollReceive(), ::performReceive() and ::update_connections().
******
Fix for bug#19559313, call getPerformStateString() with a nodeId arg instead of a state argument