Bug #19388163: PASSWORD VALIDATION PLUGIN CRASH WITH USER VARIABLE + DICTIONARY LOOKUP
The mysql_string_to_lowercase() was allocating strings for single byte csets only if the string supplied to it
was not dynamically allocated. 
But it was freeing it unconditionally via mysql_string_free(). 
This function is called by the password validator when verifying against a dictionary file.

Thus with certain arguments a double free was ocurring.

Fixed by making sure mysql_string_to_lowercase() does always allocate a new mysql string class.