WL#7165 Optimizing MTS scheduling by increasing the parallelization window on master
The patch implements the WL agenda. That includes logical timestamping
transactions on the master side with two timestamps one of which
enumerates the transaction in the binary log, and the other identifies
an earlier transaction on which the current one may depend on.

The slave applier tracks dependencies to schedule a transaction only
when the dependency parent transaction is committed.
To activate the logical clock scheduler one needs to specify

  mysql> set @@global.slave_parallel_type=logical_clock;

Values of the two timestamps aka `last_committed' and
`sequence_number' can be seen in binlog events through

  shell> mysqlbinlog binlog_file | grep 'last_committed\|sequence_number'.

The timestamp values are valid within a single binlong. They "rotate"
along with the binlog files.

A sane (recommended) mumber of workers is that that slightly exceed
the size of a parallelization window as defined as an average of
a subtract of

   sequence_number - last_committed

across transactions in the binlog.