Bug#19389709 PERFORMANCE_SCHEMA DIGESTS DO NOT HANDLE NEGATIVE NUMBERS WELL
Fix for MySQL 5.6

Before this fix, query digests computed for the performance schema
would fail to recognise literal values like "+1" or "-1" as values,
and generate a digest like "+ ?" or "- ?" instead of just "?".

This in turns caused failures to recognize list of values,
creating many different digests for queries that should have the same
signature.

The root cause is that no recognition pattern was implemented
for unary plus and unary minus operators on literals.

This fix implements the following reduces:
- <unary +> <NUM> as "?"
- <unary -> <NUM> as "?"
in the query digest computation.