MTR contains a function called fix_win_paths() which will convert '\' to '/' in any output string if it finds a path, when compiled on Windows.  This is done so that result files can be consistent on all platforms.  In order to identify a path, mysqltest sets up an array in init_win_path_patterns() of things that a path might contain.  This list contains common directories being used as well as "./test/" to catch relative paths.
WL#6205 introduces general tablespaces that exist in the datadir and thus do
not have "./test/".  They look something like "./my_tablespace.ibd". When
these are put into error messages printed to MTR result files on Windows,
they look like ".\my_tablespace.ibd".   So the path identifier list needs to
expand to include any white-space delimited word that contains ".ibd".

Along with this, the function fix_win_paths() needs to change so that when it
finds a space delimited word that looks like a path, it will back up to the
beginning of the word before replacing '\' with '/'.  This was needed because
all the previously existing path id strings occur at the start of the path,
whereas ".ibd" is at the end.

Numerous changes are made to InnoDB test files because of this change:
1) mtr.add_suppression() messages should not be written to the result file
when they contain space delimited words like "\.ibd".  If so, they would get
converted to "/.ibd" on Windows and not converted on Unix.  Wherever this is
found, change the test to keep all add_suppression messages out of the result
file.
2) These kind of lines were fixed to use "." instead of "\." in path names.
      let SEARCH_PATTERN= InnoDB: Tablespace \d+ was not found at
.*alt_dir.test.yn\.ibd\.;
3) Improve the result file comments for some tests where these problems are
found.