WL#6128, WL#6972
Follow up patch to fix a bug discovered when merging main into the
local tree.

Due to the fact that we keep an internal mapping between a number
(sidno) and UUIDs of GTIDs, there was an bug that was preventing, in
some cases, to map the correct UUID into a GTID. This was happening
because the Gtid_set, used to collect GTIDs, has its own (local) map
which is not in sync with the global map. Actually, we do not want it
to be, otherwise we be facing a synchronization penalty. The offending
behavior is that when switching the tracking mode from OFF to
OWN_GTID, and when GTIDs had already executed before this step - thus
having sidnos in the global map - we could end up with different
sidnos for the same transaction on the local map when compared to the
global map.

We fix this by using only sidnos local to the tracker when filling in
the map. At the same time, we avoid incurring into synchronization
overhead with the global map, as we are not really using it.

Added also a test case to validate this fix.

The side effect of this change is that we need to clear the local
sid_map everytime we start a new transaction, since the next one may
be tracked with ALL_GTIDS, which copies the entire executed_gtids into
the local set. This again would lead to a mismatch if the sid_map was
not pristine when this operation takes place. A consequence is that an
assertion in ensure_sidno had to be removed since it is a bit
ambiguous when sets with sid_maps that have been cleared are used.