Bug#19584181: CRASH AFTER DDL RETURNS WARNINGS - CORRUPT ERROR_HANDLER
The problem was that if a CREATE or ALTER TABLE partitioning statement
results in an ER_WRONG_TYPE_COLUMN_VALUE_ERROR error being
reported, sql_mode would by mistake be reset to 0.

This could result in corruption of the stack of Internal_error_handlers
if the ALTER TABLE statement was executed in STRICT mode.

This patch fixes the problem by making sure the sql_mode is not
changed by the CREATE/ALTER TABLE partitioning statement.