WL#6819: GCS Replication: Overview Post fix: initialization order problems
Due to the their location on the initialization process, some routines
were affected by several problems.
This included the extraction of the server id (that was absent on
server restart) or the initialization of the recovery relay logs
(that clashed with replication ones).
Adding some of the routines to the delayed initialization routine in
the plugin and making this same routine be executed later on another
hook solves these issues.