BUG#19566148 - VALGRIND ERRORS WHEN THERE IS A PORT CLASH                DURING SERVER STARTUP. Problem Description & Fix: Valgrind report errors relating to std::string in network_init function during server startup if there is a port clash. The cause under error path in network_init we call unireg_abort which calls exit c library function. network_init function some stack allocated std::string objects and this objects destructor is not invoked if exit application via exit in network_init. This causes valgrind to report errors (since gcc std::string implementation uses dynamic memory which doesn't get freed because of non-invocation of the destructors).
The fix is to make network_init return error to the
caller function and this function will invoke unireg_abort.
This will ensure that destructor functions of std::string
are called as part of the normal stack unwinding code
of libstdc++. The fix also const correct the std::string 
objects used in network_init.