WL#6842 Transaction Applier Module
(Pipeline and Reset Master support post fix)

Aiming to add better support to RESET MASTER commands, actions are
added to pipeline's handlers. These allow us to preform actions before
starting some of the handlers and also configure them in a generic
way.

About reset master support this is now based on the after reset master
hook. The hook triggers a flag change on a stopped plugin and trough
the use of pipeline actions this will cause a reset in the applier
handler when it starts.

There is also a small fix to activate the certification broadcast
thread that had no proper interfaces to function since the last biding
refactor

Besides this, there is also a fix for the display of the stable set on
situations where more than on set of GTIDs are present.

Finally, there is a test correction on
rpl_gcs_perfschema_node_status.result
with a code for pre test cleanup added.