Bug#18062698 - INNODB LOCK MONITOR CRASHES IF TABLE DROPPED AT THE RIGHT TIME ON DEBUG BUILD
When buf_page_get_with_no_latch is called in lock_print_info_all_transactions,
the space maybe deleted by drop table.
So, we need to check if the space is deleted or not before getting a
page, and prevent the space be deleted as well.

This patch also fixed bug18173184.

rb#6509 Approved by Jimmy.