This is post_push patch for WL#7165. Recommitting cause by a loose assertion, in a new function, to have been corrected now.
The patch addresses few cases that were discovered by analysis of failing
tests on PB trunk.

A. failed but still loggable DML on a non-transactional table.
In such case "@@global" max committed transaction left not
incremented.
That causes sequentially following transaction not to regard it
as its parent. Failure to do so is caught by rpl_view and
rpl_stm_EE_err2.

B. Duplicate sequence_number:s can appear as result of logging from
two caches. A traditional use case for that is mixed engine transaction.
Logging therefore is refined to timestamp separately either cache content,
and to force the 1nd (the statement cache part) be the commit parent to the 2nd.
It is important to serialize the two such way 'cos of use cases where
LOAD DATA is found out to be potentially two-cache loggable
(see rpl_parallel_load_data as example).

C. gaps in sequence_number that make the scheduler to hang
in waiting for non-existent transaction commit.
Gaps appear in circular replication and can show up in GTID
autoposition-based recovery (it all seems there's no dedicated
test for that, otherwise it would fail as well).
It's fixed with allowing "positive" (skipped groups) gaps. Negative
(as jump back in past) are made to be errored out exclusively.
A transaction that is found to succeed a gap waits for all
preceding to finish (fallback to sequential).
And at that point its dependency with *any* commit parent is satisfied.
rpl_circular_for_4_hosts verifies this solution to [ pass ].

D.There was a sporadic failure (assert) in rpl_mts_submode_switch which
turns out to be a result of sort of a race when while it's proven
all transactions are committed some Workers can still have
undeleted yet events in their private queues.
This case is specific to exceptional handling of CREATE..SELECT @user_var,
there're 2 hunks in rpl_mts_submode_switch.cc for it.

rpl_mts_slave_preserve_commit_order_error.test is blocked out from
execution with Logical-clock. It should be refined, suggested
as a separate task (todo: report a task to list all tests that
are incompatible to LC).

Finally, suppression of an expected warning in 
rpl_user_var_zero_length_name is done.

There's a separate issue exposed by rpl_wl6292 as results mismatch.
In fact it's a failure to adapt Worker THD context to requirements
carried by the master version.
Although this issue is not witnessed by PB elsewhere, not least thanks
to the test had been until recently blocked by running with Logical_clock,
Database type MTS is prone to that too.
See it as Bug #19630322 MTS: WORKERS MAY NOT ADAPT TO THE MASTER VERSION


All changes to parts A,B are located in binlog.cc. Diff file is annotated
more specifically to separate the two.
All changes to part C are located in rpl_mts_submode.cc.