BUG#18903979 - THE TABLE IS FULL ERROR WHEN RUNNING ON EXT3 WITH O_DIRECT
The fallocate() is not supported by ext3 file system. If O_DIRECT is not
specified, internal posix_fallocate() of glibc would be called instead and
things go well, if O_DIRECT is specified, internal posix_fallocate() would
fail due to alignment issue.

We could accept the failure of posix_fallocate() and just print the
error message and go on, so that the file operation can go on and we
won't be blocked in ext3 file system with O_DIRECT. If it's a
FusionIO HW/Firmware, it could be fail later, since we print the
error message already, we could know where it went wrong exactly.

rb#6349, approved by Jimmy.